NewClusterConfig and ConfigWithCluster don't make a copy of the rest config
It turns out that copying inside those functions leads to confusion.
There are parts that make an additional copy, parts that rely on the fact
that the copy is made inside, and other parts doing nothing.

Thus it would be better if callers could make a copy.
It would be explicit and would reduce the number of copies in general.

In addition to that NewClusterConfig has been renamed to SetMultiClusterRoundTripper and ConfigWithCluster to SetCluster